Fixing integration test and startup test class

+++ src/Hydra.Order.API/Models/OrderItemDto.cs
@@ -5,6 +5,6 @@ namespace Hydra.Order.API.Models
     public class OrderItemDto
     {
         public Guid Id { get; set; }
-        public decimal Qty { get; set; }
+        public int Qty { get; set; }
     }
 }
